Add --sample-services command, exposing remote sampling strategy endpoint
This way, with an invocation like

    ./jaegercat --sample-services foo-service,bar-service

the utility will expose an HTTP endpoint on 127.0.0.1:5778, and when asked
for a service's sample strategy via GET /sampling?service=foo-service, it
will respond with a hard-coded "output all samples" setting.

The mentioned endpoint is queried in that way by the jager-go library every
minute, so it takes moment (worst case: the whole minute) to have the
clients pick up that they are now supposed to output all their spans.

This can be built for Linux locally using docker:

    $ docker run -v $(pwd):/src --workdir /src -it rust:1.25.0 bash
    # rustup target add x86_64-unknown-linux-musl
    # cargo build --target x86_64-unknown-linux-musl

and then copied from (outside of the docker container)

    target/x86_64-unknown-linux-musl/debug/jaegercat

Signed-off-by: Stephan Renatus <srenatus@chef.io>